fix/Delete env, add gitignore, estrutura questoes
